Completed avx2_double_*.cxx example programs.

+++ examples/simd_stat/stat_float_arithmetic_mean.cxx
@@ -18,9 +18,9 @@ int main()
   std::int32_t n = x.size();
 
   // Compute arithmetic mean
-  double m1 = nola::arithmetic_mean_serial( n, x.data() );
-  //double m2 = nola::arithmetic_mean_avx2(   n, x.data() );
-  //double m3 = nola::arithmetic_mean_avx512( n, x.data() );
+  float m1 = nola::arithmetic_mean_serial( n, x.data() );
+  //float m2 = nola::arithmetic_mean_avx2(   n, x.data() );
+  //float m3 = nola::arithmetic_mean_avx512( n, x.data() );
 
   // Display result
   std::cout << "\narithmetic mean (serial) = " << m1 // m1 = 0.6000
